Initial Commit
Some small issues with the excel formulas need to be corrected. But the columns of xp left and xp per day should be red if not maxed, and green if maxed. The xp till max cell doesnt work for some reason. I really cant be bothered atm to look at that.